refactor(utils): extract shared frequency parsing utilities
- Create shared module src/utils/frequency-parsing.js
- Move parseFrequency, isCalendarBasedUnit, addCalendarIncrement to shared module
- Remove duplicated functions from gap-detection.js and util.js
- Fix unused loop variable in alignSeriesToLinearTimeline
- Fix redundant .map(ts => ts) operation

Addresses code review feedback from PR #445
